+++ src/Functions/AttributePointChanger.js
@@ -1,0 +1,6 @@
+export function attributePointChanger(att){
+    let rand= Math.floor(Math.random()*att);
+    return rand;
+    }
+
+    //randomise the attack between up to its maxiumum
