+++ src/Functions/RandomInt.js
@@ -1,0 +1,10 @@
+
+
+
+
+export function randomInt(){
+    let rand= Math.ceil(Math.random()*151);
+    return rand;
+    }
+
+    //returns random int for the pokemon fetch calls on the API.set to 151 for only the original and arguably best pokemon
