+++ src/heading/index.js
@@ -1,10 +1,12 @@
 import React from 'react';
 
 
-function H1({heading}){
+function H1({heading="Pokemon"}){
     return(
         <h1>{heading}</h1>
     )
 }
+
+//h1 heading for main body
 
 export default H1;
